Improved config mgmt and fixed out of bounds panic
Changed the way the config is loaded to use defined structs instead of
nested maps of interface{}.  Now, the config is loaded into a struct
during initialization and the values are referenced by the struct fields
as appropriate.  This refrains from reading and parsing the config file
twice (once when loading the default sources and again when loading any
analyzers.

Additionally, this change adds a length check to fix the panic shown in
issue #2.