Update removal of bad packages from package list.
Updated the way that bad packages are removed from the package list to
be more efficient. The previous method ran in O(n*m) time (n = num of
all packages, m = num of bad packages), and the updated method runs in
O(n) and also cleans up some of the surrounding code to make it more
readable.